Revert "Add compile support for  MSVC"
This reverts commit 5ee1ffd110b50e8e377fc6338e69bc14aa33020d.

+++ inc/timer.h
@@ -1,8 +1,3 @@
-/**
- * @file timer.h
- * @author jkilopu
- * @brief Stopwatch in mymines.
- */
 #ifndef __TIMER_H
 #define __TIMER_H
 
@@ -20,9 +15,9 @@
  */
 typedef struct _timer {
     SDL_TimerID timer_id;
-    unsigned int time_passed; ///< Time in seconds
-    unsigned int time_block_x;
-    unsigned int time_block_y;
+    unsigned short time_passed; ///< Time in seconds
+    unsigned short time_block_x;
+    unsigned short time_block_y;
 } Timer;
 
 //-------------------------------------------------------------------
@@ -30,7 +25,7 @@ typedef struct _timer {
 //-------------------------------------------------------------------
 
 void set_timer(Timer *p_timer);
-void set_timer_pos(Timer *p_timer, unsigned int win_x, unsigned int win_y);
+void set_timer_pos(Timer *p_timer, unsigned short win_x, unsigned short win_y);
 void unset_timer(Timer *p_timer);
 void draw_timer(Timer *p_timer);
 Uint32 timer_callback(Uint32 interval, void *param);
